requirements.txt
python-dateutil~=2.7.3
django~=2.0.1
django-extensions~=2.1.0
django-money~=0.14.3
django-mptt~=0.9.1
django-prices~=1.0.3
psycopg2>=2.7,<2.8 --no-binary psycopg2
pytest~=3.7.2
pytest-cov~=2.5.1
pytest-django~=3.4.1
pytest-lazy-fixture~=0.4.1
unittest2~=1.1.0
urllib3~=1.23
